requirements-dev.txt
Django==3.2.19
djangorestframework==3.14.0
django-apscheduler==0.6.2

pylint==2.17.4
PyMySQL==1.1.0
pytest==7.4.0